+++ crates/sncast/tests/e2e/multicall/run.rs
@@ -1,4 +1,5 @@
 use crate::helpers::constants::{ACCOUNT_FILE_PATH, MULTICALL_CONFIGS_DIR, URL};
+use crate::helpers::fee::apply_test_resource_bounds_flags;
 use crate::helpers::fixtures::create_and_deploy_oz_account;
 use crate::helpers::runner::runner;
 use indoc::{formatdoc, indoc};
@@ -25,12 +26,13 @@ async fn test_happy_case(account: &str) {
         "--account",
         account,
         "multicall",
+        "run",
         "--url",
         URL,
-        "run",
         "--path",
         path,
     ];
+    let args = apply_test_resource_bounds_flags(args);
 
     let snapbox = runner(&args).env("SNCAST_FORCE_SHOW_EXPLORER_LINKS", "1");
     let output = snapbox.assert();
@@ -67,12 +69,13 @@ async fn test_calldata_ids() {
         "--account",
         "my_account",
         "multicall",
+        "run",
         "--url",
         URL,
-        "run",
         "--path",
         path,
     ];
+    let args = apply_test_resource_bounds_flags(args);
 
     let snapbox = runner(&args)
         .env("SNCAST_FORCE_SHOW_EXPLORER_LINKS", "1")
@@ -105,9 +108,9 @@ async fn test_invalid_path() {
         "--account",
         "my_account",
         "multicall",
+        "run",
         "--url",
         URL,
-        "run",
         "--path",
         "non-existent",
     ];
@@ -144,9 +147,9 @@ async fn test_deploy_fail() {
         "--account",
         "my_account",
         "multicall",
+        "run",
         "--url",
         URL,
-        "run",
         "--path",
         path,
     ];
@@ -179,9 +182,9 @@ async fn test_invoke_fail() {
         "--account",
         "my_account",
         "multicall",
+        "run",
         "--url",
         URL,
-        "run",
         "--path",
         path,
     ];
@@ -214,9 +217,9 @@ async fn test_deploy_success_invoke_fails() {
         "--account",
         "my_account",
         "multicall",
+        "run",
         "--url",
         URL,
-        "run",
         "--path",
         path,
     ];
@@ -249,12 +252,13 @@ async fn test_numeric_inputs() {
         "--account",
         "my_account",
         "multicall",
+        "run",
         "--url",
         URL,
-        "run",
         "--path",
         path,
     ];
+    let args = apply_test_resource_bounds_flags(args);
 
     let snapbox = runner(&args)
         .env("SNCAST_FORCE_SHOW_EXPLORER_LINKS", "1")