+++ CHANGELOG.md
@@ -11,16 +11,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 - `--max-threads` flag to control the maximum number of threads used for test execution
+- `sncast multicall execute`, which allows to pass calls directly as CLI arguments. Read more [here](https://foundry-rs.github.io/starknet-foundry/starknet/multicall.html#multicall-with-cli-arguments)
 
 #### Fixed
 
 - Bug with invalid function name mappings for functions with `#[test]` attribute
+- `--exit-first` flag now correctly stops execution for all packages when tests are run in a workspace
 
 ### Cast
 
 #### Changed
 
-- Moved RPC and fee arguments (e.g., `--url`) from the `run` subcommand to the parent `multicall` command
+- `sncast balance` and `sncast tx-status` commands have been moved under `sncast get` subcommand (`sncast get balance`, `sncast get tx-status`). The old commands still work, but will be removed in the future.
+- Referencing an id in multicall configuration files now requires the `@` prefix.
 
 ## [0.57.0] - 2026-02-24